Add django-migration-slow-default rule
This rule checks for Django migrations that might be really slow in
production, as they add a NOT NULL column to an existing table, and
since Django does not leverage SQL DEFAULT it has backfill all rows with
the default value row by row.

The correct approach in most cases is to write a manual `runSQL`
migration which adds the column by hand with an SQL default and
explicitly passes the state_operations argument so that Django doesn't
complain about code and database being out of sync.