Add semgrep rule on safe field removal
RemoveField operations should span across two versions:

* In the first version, the field is only removed from the state (as
  Django calls it), meaning that it is no longer accessible from Django
  code but still exists in the database, this guarantees that if a
  rollback happens during deployment the data still exists and can be
  used by the code running in production at the time of deployment.

* In the subsequent version, the column can be safely removed from the
  database, as neither the new version nor the previous version use said
  table.

If the field being removed has a NOT NULL constraint, this constraint
should be dropped before the RemoveField state-only operation is called,
or alternatively given an SQL default value, otherwise the new version
will try to create rows without providing a value for the NOT NULL
column.

This commit adds a semgrep rule that tries to guarantee that RemoveField
operations are safe according to the preceding guidelines, semgrep
cannot however tell whether the field being removed has a NOT NULL
constraint or not, in this case the safety is deferred to the individual
developer who will need to write the manual migration or possibly to
helpers that already do the constraint removal.