Add semgrep rule on RenameField migration operations
This commit adds a rule that checks for RenameField migration
operations.

These operations are dangerous because while version N of the django
application will work fine, version N-1 will not be able to find the
column that it uses for the field that was renamed.

The easiest solution in these scenarios is to rename the field in the
code, but add to the field definition an explicit db_column attribute
with the name of the old field, this means that while the field is
renamed in the code and thus in Django, the database is not impacted and
both version N and version N-1 can function despite the changes.

Note: technically speaking there is a database change but that
ultimately has no impact, this is due to a limitation in django, see
https://code.djangoproject.com/ticket/24260

Given said limitation, it is safest to wrap the FieldRename and
AlterField migration operations inside a SeparateDatabaseAndState
migration operation, which leads to no SQL output during the migration.